+++ daemon/cmd/daemon/monitor_providers.go
@@ -6,13 +6,6 @@ import (
 	"waypaper-engine/daemon/internal/monitor"
 )
 
-// defaultMonitorProviders builds the standard monitor-provider chain used by
-// both the running daemon and the `monitors --direct` diagnostic. v may be
-// nil when called from `--direct` without a loaded config — utauri falls back
-// to its default control-socket path in that case.
-//
-// Evaluation order on Wayland (by priority): native zwlr_output_management first,
-// then wal-qt HTTP status (subset topology), then legacy randr-style probes.
 func defaultMonitorProviders(v backend.ConfigReader) []monitor.MonitorProvider {
 	return []monitor.MonitorProvider{
 		monitor.NewWaylandProvider(),
