+++ daemon/internal/handler/backendshandler/activate_backend_response.go
@@ -1,6 +1,5 @@
 package backendshandler
 
-// ActivateBackendResponse is the JSON body for POST /backends/{name}/activate.
 type ActivateBackendResponse struct {
 	Status  string `json:"status"`
 	Backend string `json:"backend"`
