+++ daemon/internal/handler/backendshandler/backends.go
@@ -2,21 +2,18 @@ package backendshandler
 
 import (
 	"net/http"
-
-	"github.com/go-chi/chi/v5"
-
 	"waypaper-engine/daemon/internal/backend"
 	"waypaper-engine/daemon/internal/control"
 	"waypaper-engine/daemon/internal/handler/httpjson"
+
+	"github.com/go-chi/chi/v5"
 )
 
-// BackendHandler handles /backends endpoints.
 type BackendHandler struct {
 	registry backend.Registry
 	control  *control.Controller
 }
 
-// NewBackendHandler creates a BackendHandler.
 func NewBackendHandler(registry backend.Registry, control *control.Controller) *BackendHandler {
 	return &BackendHandler{
 		registry: registry,
@@ -24,12 +21,10 @@ func NewBackendHandler(registry backend.Registry, control *control.Controller) *
 	}
 }
 
-// List handles GET /backends.
 func (h *BackendHandler) List(w http.ResponseWriter, r *http.Request) {
 	httpjson.WriteJSON(w, http.StatusOK, h.registry.Available())
 }
 
-// Activate handles POST /backends/{name}/activate.
 func (h *BackendHandler) Activate(w http.ResponseWriter, r *http.Request) {
 	name := chi.URLParam(r, "name")
 	result, err := h.control.ActivateBackend(r.Context(), name)
