+++ daemon/internal/handler/confighandler/config.go
@@ -5,24 +5,20 @@ import (
 	"errors"
 	"net/http"
 	"net/url"
-
-	"github.com/go-chi/chi/v5"
-
 	"waypaper-engine/daemon/internal/control"
 	"waypaper-engine/daemon/internal/handler/httpjson"
+
+	"github.com/go-chi/chi/v5"
 )
 
-// ConfigHandler handles /config endpoints.
 type ConfigHandler struct {
 	control *control.Controller
 }
 
-// NewConfigHandler creates a ConfigHandler.
 func NewConfigHandler(ctrl *control.Controller) *ConfigHandler {
 	return &ConfigHandler{control: ctrl}
 }
 
-// GetConfig handles GET /config.
 func (h *ConfigHandler) GetConfig(w http.ResponseWriter, r *http.Request) {
 	merged, err := h.control.MergedConfigJSON()
 	if err != nil {
@@ -33,7 +29,6 @@ func (h *ConfigHandler) GetConfig(w http.ResponseWriter, r *http.Request) {
 	httpjson.WriteJSON(w, http.StatusOK, merged)
 }
 
-// PatchConfig handles PATCH /config.
 func (h *ConfigHandler) PatchConfig(w http.ResponseWriter, r *http.Request) {
 	var body map[string]map[string]any
 	if err := httpjson.ParseBody(r, &body); err != nil {
@@ -57,7 +52,6 @@ func (h *ConfigHandler) PatchConfig(w http.ResponseWriter, r *http.Request) {
 	httpjson.WriteJSON(w, http.StatusOK, merged)
 }
 
-// PostResetAll handles POST /config/reset — factory-reset every section and backend subtree.
 func (h *ConfigHandler) PostResetAll(w http.ResponseWriter, r *http.Request) {
 	if err := h.control.ResetAllConfigToDefaults(r.Context()); err != nil {
 		httpjson.WriteError(w, http.StatusInternalServerError, err.Error())
@@ -71,7 +65,6 @@ func (h *ConfigHandler) PostResetAll(w http.ResponseWriter, r *http.Request) {
 	httpjson.WriteJSON(w, http.StatusOK, merged)
 }
 
-// GetSection handles GET /config/{section}.
 func (h *ConfigHandler) GetSection(w http.ResponseWriter, r *http.Request) {
 	section := chi.URLParam(r, "section")
 	if section == "backend" {
@@ -88,7 +81,6 @@ func (h *ConfigHandler) GetSection(w http.ResponseWriter, r *http.Request) {
 	httpjson.WriteJSON(w, http.StatusOK, data)
 }
 
-// PatchSection handles PATCH /config/{section}.
 func (h *ConfigHandler) PatchSection(w http.ResponseWriter, r *http.Request) {
 	section := chi.URLParam(r, "section")
 	if section == "backend" {
@@ -98,7 +90,6 @@ func (h *ConfigHandler) PatchSection(w http.ResponseWriter, r *http.Request) {
 	h.patchSection(w, r, section)
 }
 
-// GetNamedBackendConfig handles GET /config/backends/{backend}.
 func (h *ConfigHandler) GetNamedBackendConfig(w http.ResponseWriter, r *http.Request) {
 	name := namedBackendFromRequest(r)
 	if name == "" {
@@ -119,7 +110,6 @@ func (h *ConfigHandler) GetNamedBackendConfig(w http.ResponseWriter, r *http.Req
 	_, _ = w.Write(raw)
 }
 
-// PatchNamedBackendConfig handles PATCH /config/backends/{backend}.
 func (h *ConfigHandler) PatchNamedBackendConfig(w http.ResponseWriter, r *http.Request) {
 	name := namedBackendFromRequest(r)
 	if name == "" {
@@ -146,7 +136,6 @@ func (h *ConfigHandler) PatchNamedBackendConfig(w http.ResponseWriter, r *http.R
 	httpjson.WriteJSON(w, http.StatusOK, map[string]string{"status": "updated"})
 }
 
-// PostResetNamedBackendConfig handles POST /config/backends/{backend}/reset.
 func (h *ConfigHandler) PostResetNamedBackendConfig(w http.ResponseWriter, r *http.Request) {
 	name := namedBackendFromRequest(r)
 	if name == "" {
