--- daemon/internal/handler/foldershandler/delete_folder_response.go
@@ -1,7 +1,0 @@
-package foldershandler
-
-// DeleteFolderResponse is the JSON body for DELETE /folders/{id}.
-type DeleteFolderResponse struct {
-	Deleted bool   `json:"deleted"`
-	Mode    string `json:"mode"`
-}
