--- daemon/internal/handler/foldershandler/move_images_response.go
@@ -1,6 +1,0 @@
-package foldershandler
-
-// MoveImagesResponse is the JSON body for POST /folders/move-images.
-type MoveImagesResponse struct {
-	Moved int `json:"moved"`
-}
