+++ daemon/internal/handler/healthhandler/capabilities_response.go
@@ -1,6 +1,5 @@
 package healthhandler
 
-// CapabilitiesResponse is the JSON body for GET /capabilities.
 type CapabilitiesResponse struct {
 	FfmpegAvailable bool `json:"ffmpeg_available"`
 }
