+++ daemon/internal/handler/healthhandler/healthz_response.go
@@ -1,6 +1,5 @@
 package healthhandler
 
-// HealthzResponse is the JSON body for GET /healthz.
 type HealthzResponse struct {
 	Status               string   `json:"status"`
 	MonitorStackVersion  int      `json:"monitor_stack_version"`
